cleaned up code, got the lm to do better with straight-up variables
@danielhong98 @maxwagner . I just updated the code. HEre's a summary of
the changes:

* Cleaned up the code a bit so that it is easier to read
* Got the lm to work better off the untransformed variables. The trick
is to cancel the intercept, which makes sense, if you do nothing, then
you're not going to score any wins :). However, I'm not thrilled about
the residual distribution charts, it looks like transformations are
needed after all so that we don't violate the assumptions. Having said
that, this model is very easy to interpret, perhaps we can use it to
explain the variables/influencers but rely on the advanced models for
predictions.

Anyways, I leave you with that. Let me know your thoughs.

Best,